loading/lazyload/change_attribute_to_off_iframe.tentative.sub.html

<!DOCTYPE html>
<head>
  <title>Iframes load immediately when lazyload attribute is changed to "off"</title>
  <link rel="author" title="Scott Little" href="mailto:sclittle@chromium.org">
  <link rel="help" href="https://github.com/whatwg/html/pull/3752">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<body>
  <div style="height:10000px;"></div>
  <iframe id="attribute_unset" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_1.html" width="200px" height="100px"></iframe>
  <iframe id="attribute_invalid" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_2.html" width="200px" height="100px" lazyload="invalid_value"></iframe>
  <iframe id="attribute_auto" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_3.html" width="200px" height="100px" lazyload="auto"></iframe>
  <iframe id="attribute_on" src="https://{{hosts[][www]}}:{{ports[https][0]}}/loading/lazyload/resources/subframe_4.html" width="200px" height="100px" lazyload="on"></iframe>
</body>

<script>
  var t = async_test("Test that below viewport iframes that have their lazyload attributes changed to off are loaded.");

  var has_attribute_unset_loaded = false;
  var has_attribute_invalid_loaded = false;
  var has_attribute_auto_loaded = false;
  var has_attribute_on_loaded = false;

  function checkIfDone() {
    if (has_attribute_unset_loaded &&
        has_attribute_invalid_loaded &&
        has_attribute_auto_loaded &&
        has_attribute_on_loaded) {
      t.done();
    }
  }

  document.getElementById("attribute_unset").addEventListener("load", t.step_func(function() {
    assert_false(has_attribute_unset_loaded);
    has_attribute_unset_loaded = true;
    checkIfDone();
  }));
  document.getElementById("attribute_invalid").addEventListener("load", t.step_func(function() {
    assert_false(has_attribute_invalid_loaded);
    has_attribute_invalid_loaded = true;
    checkIfDone();
  }));
  document.getElementById("attribute_auto").addEventListener("load", t.step_func(function() {
    assert_false(has_attribute_auto_loaded);
    has_attribute_auto_loaded = true;
    checkIfDone();
  }));
  document.getElementById("attribute_on").addEventListener("load", t.step_func(function() {
    assert_false(has_attribute_on_loaded);
    has_attribute_on_loaded = true;
    checkIfDone();
  }));

  window.addEventListener("load", t.step_func_done(function() {
    document.getElementById("attribute_unset").setAttribute("lazyload", "off");
    document.getElementById("attribute_invalid").setAttribute("lazyload", "off");
    document.getElementById("attribute_auto").setAttribute("lazyload", "off");
    document.getElementById("attribute_on").setAttribute("lazyload", "off");
  }));
</script>
